requirements.txt
tensorflow==2.13.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
django>=4.0.0
djangorestframework>=3.12.0 